I removed your int() again, because it was hiding the cause instead of fixing it.

+++ nt/transform/module_stft.py
@@ -132,7 +132,7 @@ def _stft_frames_to_samples(frames, size, shift):
     :param shift: Hop in samples.
     :return: Number of samples in time domain.
     """
-    return int(frames * shift + size - shift)
+    return frames * shift + size - shift
 
 
 def _biorthogonal_window_loopy(analysis_window, shift):
